requirements.txt
discord.py
jishaku
dblpy
aiosqlite
